+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+* 2.0.0 (???)
+
+  * [#21] Add configurable timeout for single optimizer. Use
+    symfony-process component as an implementation.
+
 * 1.2.2 (2019-06-21)
 
   * [#58] Easy way to add custom optimizers.